+++ src/cmd/compile/internal/ssa/rewriteARM.go
@@ -15120,7 +15120,7 @@ func rewriteValueARM_OpMove(v *Value) bool {
 	}
 	// match: (Move [s] {t} dst src mem)
 	// cond: (s > 512 || config.noDuffDevice) || t.(Type).Alignment()%4 != 0
-	// result: (LoweredMove [t.(Type).Alignment()] 		dst 		src 		(ADDconst <src.Type> src [s-moveSize(t.(Type).Alignment(), config)]) 		mem)
+	// result: (LoweredMove [int64(t.(Type).Alignment())] 		dst 		src 		(ADDconst <src.Type> src [s-moveSize(t.(Type).Alignment(), config)]) 		mem)
 	for {
 		s := v.AuxInt
 		t := v.Aux
@@ -15131,7 +15131,7 @@ func rewriteValueARM_OpMove(v *Value) bool {
 			break
 		}
 		v.reset(OpARMLoweredMove)
-		v.AuxInt = t.(Type).Alignment()
+		v.AuxInt = int64(t.(Type).Alignment())
 		v.AddArg(dst)
 		v.AddArg(src)
 		v0 := b.NewValue0(v.Pos, OpARMADDconst, src.Type)
@@ -16942,7 +16942,7 @@ func rewriteValueARM_OpZero(v *Value) bool {
 	}
 	// match: (Zero [s] {t} ptr mem)
 	// cond: (s > 512 || config.noDuffDevice) || t.(Type).Alignment()%4 != 0
-	// result: (LoweredZero [t.(Type).Alignment()] 		ptr 		(ADDconst <ptr.Type> ptr [s-moveSize(t.(Type).Alignment(), config)]) 		(MOVWconst [0]) 		mem)
+	// result: (LoweredZero [int64(t.(Type).Alignment())] 		ptr 		(ADDconst <ptr.Type> ptr [s-moveSize(t.(Type).Alignment(), config)]) 		(MOVWconst [0]) 		mem)
 	for {
 		s := v.AuxInt
 		t := v.Aux
@@ -16952,7 +16952,7 @@ func rewriteValueARM_OpZero(v *Value) bool {
 			break
 		}
 		v.reset(OpARMLoweredZero)
-		v.AuxInt = t.(Type).Alignment()
+		v.AuxInt = int64(t.(Type).Alignment())
 		v.AddArg(ptr)
 		v0 := b.NewValue0(v.Pos, OpARMADDconst, ptr.Type)
 		v0.AuxInt = s - moveSize(t.(Type).Alignment(), config)
